requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
